# src/nswag/Helmer.PetStore.Nswag.Api/Models/User.cs
namespace Helmer.PetStore.Nswag.Api.Models;

public class User
{
    public long? Id { get; set; }
    public string? Username { get; set; }
    public string? FirstName { get; set; }
    public string? LastName { get; set; }
    public string? Email { get; set; }
    public string? Password { get; set; }
    public string? Phone { get; set; }

    /// <summary>User Status</summary>
    public int? UserStatus { get; set; }
}